support per-CPU map types
Per-CPU maps return an array of values, one per possible CPU, instead of a single value.
Additionally, these values are padded to multiples of 8 bytes. Add some specialised
unmarshaling code which makes dealing with these nicer.

+++ syscalls.go
@@ -158,11 +158,11 @@ func pinObject(fileName string, fd uint32) error {
 	return bpfErrNo(errNo)
 }
 
-func getObject(fileName string) (uintptr, error) {
+func getObject(fileName string) (uint32, error) {
 	ptr, errNo := bpfCall(_ObjGet, unsafe.Pointer(&pinObjAttr{
 		fileName: newPtr(unsafe.Pointer(&[]byte(fileName)[0])),
 	}), 16)
-	return ptr, bpfErrNo(errNo)
+	return uint32(ptr), bpfErrNo(errNo)
 }
 
 func getObjectInfoByFD(fd uint32, info unsafe.Pointer, size uintptr) error {
